Moved the value logic into a seperate moduel
